+++ tmf630-toolkit-attribute-filtering-autoconfigure/src/test/java/org/opentmf/query/tmf630/filtering/it/sqlnested/SqlOrderRepository.java
@@ -1,0 +1,5 @@
+package org.opentmf.query.tmf630.filtering.it.sqlnested;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SqlOrderRepository extends JpaRepository<SqlOrderEntity, Long> {}